google test x64 also works

+++ sources/libraries/msbuild.cpp
@@ -1,11 +1,12 @@
 #include <filesystem>
 #include <fstream>
+#include <algorithm>
 
 using namespace std::experimental;
 
 void msbuild_change_runtime(const filesystem::path& path_msvc)
 {
-
+	
 	for(auto& file : filesystem::directory_iterator(path_msvc))
 	{
 		if(".vcxproj" == file.path().extension())
@@ -49,6 +50,65 @@ void msbuild_change_runtime(const filesystem::path& path_msvc)
 			filesystem::copy_file(
 				path_new, 
 				path_old, 
+				filesystem::copy_options::overwrite_existing
+			);
+			filesystem::remove(path_new);
+
+		}
+
+	}
+
+}
+
+void msbuild_convert_to_x64(const filesystem::path& path_msvc)
+{
+
+	for(auto& file : filesystem::directory_iterator(path_msvc))
+	{
+
+		if(".vcxproj" == file.path().extension() || ".sln" == file.path().extension())
+		{
+
+			auto path_old(file.path());
+			auto path_new(path_old);
+			path_new.replace_extension(".tmp");
+
+			std::ifstream is(path_old.string());
+			std::ofstream os(path_new.string());
+
+			std::string buffer(1024, '\0');
+			for(
+				is.getline(&buffer[0], buffer.size());
+				is;
+				is.getline(&buffer[0], buffer.size())
+				)
+			{
+
+				std::string line(buffer.c_str());
+
+				std::string::size_type pos = 0u;
+				while(std::string::npos  != (pos = line.find("Win32", pos)))
+				{
+					line.replace(pos, 5, "x64");
+					pos += 3;
+				}
+				pos = 0u;
+				while(std::string::npos != (pos = line.find("x86", pos)))
+				{
+					line.replace(pos, 3, "x64");
+					pos += 3;
+				}
+
+				os << line << "\n";
+
+			}
+
+			os.close();
+			is.close();
+
+			filesystem::copy_file(
+				path_new,
+				path_old,
 				filesystem::copy_options::overwrite_existing
 			);
 			filesystem::remove(path_new);
